pyproject.toml
[tool.poetry]
name = "nasa-exoplanet-query"
version = "0.4.0"
description = ""
authors = ["Elicácio C. de Farias <elicaciocardoso@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
psycopg2-binary = "^2.8.6"
dynaconf = "^3.1.4"
pandas = "^1.2.5"
httpx = "^0.18.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.9.2"
isort = "^5.8.0"
black = "^21.6b0"
pytest-django = "^4.4.0"

[tool.black]
line-length = 88

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
